requirements/base.txt
asgiref==3.4.1
coverage==5.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-import-export==2.5.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
